requirements.txt
pandas
watchdog
pytest
pytest-cov
python-dotenv
pytest==8.0.0
psycopg2-binary
